testing and refactoring
