+++ index.html
@@ -24,47 +24,9 @@
 	</header>
 	<section>
 		<main>
-			<div id="main-content">
-        <div class="card" style="width: 18.5em;">
-          <div id="card-header" class="card-body">
-            <img src="../images-g/profilePic.jpg" id="profile-pic" class="card-img-top" alt="User profile picture" width="260"
-              height="260">
-            <div class="card-title">
-              <h1 id="full-name">Greg Markus</h1>
-              <h1 id="username">gregGroks13</h1>
-            </div>
-          </div>
-          <div id="main-body">
-            <div id="btn-container">
-              <button id="follow-btn">Follow</button>
-              <button id="sponsor-btn">Sponsor</button>
-              <button id="dots-btn">...</button>
-            </div>
-            <p class="card-text">I am new to tech and very willing to learn.</p>
-            <div id="follow-count">
-              <div class="follow-style">
-                <a href="#" class="follow">127 followers</a>
-                <a href="#" class="follow">30 following</a>
-              </div>
-            </div>
-            <ul id="social-info">
-              <li>Nashville, Tennessee</li>
-              <li>gregGroks13@gmail.com</li>
-              <li>https://gregGroks.com</li>
-              <li>@gregGroks</li>
-              <li>@gregGroks</li>
-            </ul>
-          </div>
-          <div id="sponsor">
-            <h1 class="h1-sm">Sponsors</h1>
-            <div id="sponsor-profile-pic"></div>
-          </div>
-          <div id="organization">
-            <h1 class="h1-sm">Organizations</h1>
-            <div id="org-profile-pic"></div>
-          </div>
-        </div>
+      <div id="main-content">
       </div>
+    </main>
 			<div class="overview-content">
 				<div class="pinned-repos"></div>
 				<div id="pinned-repo-form">
