+++ pages/packages.html
@@ -160,6 +160,6 @@
       integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
       crossorigin="anonymous"
     ></script>
-    <script type="module" src="../main.js"></script>
+    <script type="module" src="../packages.js"></script>
   </body>
 </html>
