+++ pages/repo.html
@@ -160,6 +160,6 @@
       integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
       crossorigin="anonymous"
     ></script>
-    <script src="main.js"></script>
+    <script src="../repo.js"></script>
   </body>
 </html>
